installer: extract chart rendering away from installation
Calling NewInstaller and having it create a bunch of side effects is
surprising behavior, so let's move the side effects away into their own
little func, and pass the result of that to NewInstaller.
